pyproject.toml
[tool.poetry]
name = "people"
version = "0.1.0"
description = "People service"
authors = ["Collin Mutembei <collin@solublecode.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
beanie = "^1.23.0"
fastapi-users = {version = "^12.1.2", extras = ["redis", "beanie", "oauth"]}
loguru = "^0.7.2"
python-decouple = "^3.8"
phonenumbers = "^8.13.23"
uvicorn = {version = "^0.23.2", extras = ["standard"]}
python-magic = "^0.4.27"
pandas = ">=2.0.0"
minio = "^7.1.17"
fastapi-mail = "^1.4.1"
pydantic-settings = "^2.0.3"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
ipython = "^8.16.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
